requirements.txt
SpeechRecognition==3.8.1
pydub==0.25.1
python-docx==0.8.11
